+++ govtool/frontend/src/pages/DRepDirectoryContent.tsx
@@ -14,7 +14,12 @@ import {
 } from "@hooks";
 import { DataActionsBar, EmptyStateDrepDirectory } from "@molecules";
 import { AutomatedVotingOptions, DRepCard } from "@organisms";
-import { correctVoteAdaFormat, isSameDRep, uniqBy, parseBoolean } from "@utils";
+import {
+  isSameDRep,
+  uniqBy,
+  parseBoolean,
+  correctDRepDirectoryFormat,
+} from "@utils";
 import { DRepData, DRepListSort, DRepStatus } from "@models";
 import {
   AutomatedVotingOptionCurrentDelegation,
@@ -102,7 +107,7 @@ export const DRepDirectoryContent: FC<DRepDirectoryContentProps> = ({
     return <Loader />;
   }
 
-  const ada = correctVoteAdaFormat(votingPower);
+  const ada = correctDRepDirectoryFormat(votingPower);
 
   const filteredDoNotListDReps = uniqBy(
     dRepList?.filter((dRep) => {