+++ tests/govtool-frontend/playwright/tests/3-drep-registration/dRepRegistration.dRep.spec.ts
@@ -30,25 +30,29 @@ test.describe("Logged in DReps", () => {
     await page.goto("/");
 
     await expect(page.getByTestId("voting-power-chips")).toBeVisible({
-      timeout: 20_000,
+      timeout: 60_000,
     });
 
     await expect(
       page.getByTestId("dRep-id-display-card-dashboard")
-    ).toContainText(dRep01Wallet.dRepId, { timeout: 20_000 });
+    ).toContainText(dRep01Wallet.dRepId, { timeout: 60_000 });
 
     const governanceActionsPage = new GovernanceActionsPage(page);
 
     await governanceActionsPage.goto();
 
-    await expect(page.getByText(/info action/i).first()).toBeVisible();
+    await expect(page.getByText(/info action/i).first()).toBeVisible({
+      timeout: 60_000,
+    });
 
     const governanceActionDetailsPage =
       await governanceActionsPage.viewFirstProposalByGovernanceAction(
         GovernanceActionType.InfoAction
       );
 
-    await expect(governanceActionDetailsPage.voteBtn).toBeVisible();
+    await expect(governanceActionDetailsPage.voteBtn).toBeVisible({
+      timeout: 60_000,
+    });
   });
 
   test("3H. Should Update DRep data", async ({ page }, testInfo) => {