+++ govtool/frontend/.storybook/preview.tsx
@@ -11,6 +11,7 @@ import { ModalProvider } from "../src/context/modal";
 import { CardanoProvider } from "../src/context/wallet";
 import i18n from "../src/i18n";
 import { theme } from "../src/theme";
+import { AdaHandleProvider } from "../src/context/adaHandle";
 
 const queryClient = new QueryClient();
 
@@ -29,32 +30,34 @@ const preview: Preview = {
       <QueryClientProvider client={queryClient}>
         <AppContextProvider>
           <FeatureFlagProvider>
-            <ThemeProvider theme={theme}>
-              <CardanoProvider>
-                <ModalProvider>
-                  <I18nextProvider i18n={i18n}>
-                    <MemoryRouter>
-                      <Routes>
-                        <Route
-                          path="/*"
-                          element={
-                            <div
-                              style={{
-                                margin: "0px",
-                                padding: "0px",
-                                position: "relative",
-                              }}
-                            >
-                              <Story />
-                            </div>
-                          }
-                        />
-                      </Routes>
-                    </MemoryRouter>
-                  </I18nextProvider>
-                </ModalProvider>
-              </CardanoProvider>
-            </ThemeProvider>
+            <AdaHandleProvider>
+              <ThemeProvider theme={theme}>
+                <CardanoProvider>
+                  <ModalProvider>
+                    <I18nextProvider i18n={i18n}>
+                      <MemoryRouter>
+                        <Routes>
+                          <Route
+                            path="/*"
+                            element={
+                              <div
+                                style={{
+                                  margin: "0px",
+                                  padding: "0px",
+                                  position: "relative",
+                                }}
+                              >
+                                <Story />
+                              </div>
+                            }
+                          />
+                        </Routes>
+                      </MemoryRouter>
+                    </I18nextProvider>
+                  </ModalProvider>
+                </CardanoProvider>
+              </ThemeProvider>
+            </AdaHandleProvider>
           </FeatureFlagProvider>
         </AppContextProvider>
       </QueryClientProvider>